Create tests for project controller

+++ app/controllers/api/projects_controller.rb
@@ -8,7 +8,7 @@ module Api
       @projects = Project.where("user_id = ?", current_user[:id])
       authorize @projects
       render json: { data: @projects, user: current_user }
-    end 
+    end
 
     def new
       @project = Project.new
@@ -20,45 +20,52 @@ module Api
       if @project.save
         render json: { data: @project }, status: 201
       else
-        render json: {"error": @project.errors}, status: 400
+        render json: { "error": @project.errors }, status: 400
       end
     end
 
-    def show 
+    def show
       @project = Project.find_by(id: params[:id])
       if @project
         authorize @project
         render json: { data: @project }
       else
-        render json: { "error": "Project not found"}, status: 404
+        render json: { "error": "Project not found" }, status: 404
       end
     end
 
     def update
       @project = Project.find_by(id: params[:id])
-      authorize @project
-      if @project.update_attributes(project_params)
-        render json: { data: @project}, status: 200
+      if @project
+        authorize @project
+        if @project.update_attributes(project_params)
+          render json: { data: @project }, status: 200
+        else
+          render json: { "error": @project.errors }, status: 400
+        end
       else
-        render json: {"error": @project.errors}, status: 400
+        render json: { "error": "Project not found" }, status: 404
       end
     end
 
     def destroy
       @project = Project.find_by(id: params[:id])
-      authorize @project
-      if @project.destroy
-        render json: { success: { text: "Project was successfully deleted"}}, status: 200
+      if @project
+        authorize @project
+        if @project.destroy
+          render json: { success: { text: "Project was successfully deleted" } }, status: 200
+        else
+          render json: { "error": @project.errors }, status: 400
+        end
       else
-        render json: {"error": @project.errors}, status: 400
+        render json: { "error": "Project not found" }, status: 404
       end
     end
 
     private
 
     def project_params
-      params.require(:project).permit(:name)
+      params.permit(:name)
     end
-
   end
 end
